v1

+++ lab/mod08/lab04/setup_linux.sh
@@ -5,6 +5,8 @@
 rpm --import https://packages.microsoft.com/keys/microsoft.asc
 
 
+if [ ! -f /etc/yum.repos.d/azure-cli.repo ]
+then 
 
 cat <<EOF > /etc/yum.repos.d/azure-cli.repo
 [azure-cli]
@@ -14,7 +16,10 @@ enabled=1
 gpgcheck=1
 gpgkey=https://packages.microsoft.com/keys/microsoft.asc
 EOF
-  
+fi 
+
+if [ ! -f /etc/yum.repos.d/kubernetes.repo ]
+then 
 
 cat <<EOF > /etc/yum.repos.d/kubernetes.repo
 [kubernetes]
@@ -26,7 +31,13 @@ repo_gpgcheck=1
 gpgkey=https://packages.cloud.google.com/yum/doc/yum-key.gpg https://packages.cloud.google.com/yum/doc/rpm-package-key.gpg
 EOF
 
+fi 
+
+if [ ! -f /etc/yum.repos.d/mssql-server.rep  ]
+then
 curl -o /etc/yum.repos.d/mssql-server.repo https://packages.microsoft.com/config/rhel/8/prodAZDATA.repo
+fi 
+
 
 yum install azure-cli azdata-cli kubectl -y 
 if [ $? -ne 0 ]
@@ -40,8 +51,16 @@ sleep 1
 clear 
 echo 'Installation of az cli , kubectl and azdata seems OK'
 
-kubectl version --client
 
-az --version
+for util in az azdata kubectl curl wget
+    do
+    echo Verifying $util is in path & which $util 1>/dev/null 2>/dev/null || (echo Unable to locate $util && exit 1)
+done
 
-azdata --version 
+
+TMPSIZE=$(df -h | grep tmplv | awk '{print $2}')
+
+if [ $TMPSIZE != "4.0G"  ]
+ then 
+ lvextend -L +2G -n /dev/rootvg/tmplv -r &> /dev/null 
+ fi
